+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 
 ### Changed
 
+- [#6442](https://github.com/ChainSafe/forest/issues/6442): `forest-wallet sign` and `forest-wallet verify` now apply the FRC-0102 signing envelope to the message by default. Pass `--raw` on both sides to reproduce the previous raw-bytes behaviour.
+
 ### Removed
 
 ### Fixed